[native] Structural changes to native runners and renaming
* Remove inheritance between HiveExternalWorkerQueryRunner and
  AbstractNativeRunner
* Rename AbstractNativeRunner to NativeQueryRunnerUtils
* Rename HiveExternalWorkerQueryRunner to PrestoNativeQueryRunnerUtils
