+++ presto-native-execution/src/test/java/com/facebook/presto/spark/PrestoSparkNativeQueryRunner.java
@@ -15,7 +15,7 @@ package com.facebook.presto.spark;
 
 import com.facebook.presto.hive.metastore.Database;
 import com.facebook.presto.hive.metastore.ExtendedHiveMetastore;
-import com.facebook.presto.nativeworker.AbstractNativeRunner;
+import com.facebook.presto.nativeworker.NativeQueryRunnerUtils;
 import com.facebook.presto.spi.security.PrincipalType;
 import com.google.common.collect.ImmutableList;
 import com.google.common.collect.ImmutableMap;
@@ -25,11 +25,14 @@ import java.nio.file.Path;
 import java.util.Map;
 import java.util.Optional;
 
+import static com.facebook.presto.nativeworker.NativeQueryRunnerUtils.getNativeWorkerHiveProperties;
+import static com.facebook.presto.nativeworker.NativeQueryRunnerUtils.getNativeWorkerSystemProperties;
 import static com.facebook.presto.spark.PrestoSparkQueryRunner.METASTORE_CONTEXT;
 
 public class PrestoSparkNativeQueryRunner
-        extends AbstractNativeRunner
 {
+    private PrestoSparkNativeQueryRunner() {}
+
     private static final int AVAILABLE_CPU_COUNT = 4;
 
     public static PrestoSparkQueryRunner createPrestoSparkNativeQueryRunner(
